refactor(23-01): extract OnboardingComponents with shared button factory
Move createStyledButton and shared button color constants out of
PremiumOnboardingPanel into a new top-level OnboardingComponents.kt so
the free wizard can reuse the same accent/secondary button styling.
Card painting stays panel-local since free and premium card visuals
share almost no code.
